AddNodeState 施工中

+++ 102598005_ERD/GraphicsRelation.cpp
@@ -1,7 +1,7 @@
 #include "GraphicsRelation.h"
 
 const int OFFSET = 5;
-const float SCALE = 0.8;
+const double SCALE = 0.8;
 
 GraphicsRelation::GraphicsRelation()
 {
